# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Webstore Customer API",
        "description": "Dokumentasi endpoint API customer untuk autentikasi dan fitur terkait dashboard.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Default API server"
        }
    ],
    "paths": {
...
        "/api/checkout": {
            "get": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Data checkout customer",
                "operationId": "b70816fdbd51ba6191a297cd622a1e14",
                "responses": {
                    "200": {
                        "description": "Data checkout berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Data checkout berhasil diambil.",
                                    "data": {
                                        "items": [
                                            {
                                                "id": 1,
                                                "name": "Produk A",
                                                "qty": 2,
                                                "row_total": 250000
                                            }
                                        ],
                                        "cart": {
                                            "subtotal": 250000,
                                            "shipping": 15000,
                                            "total": 265000
                                        },
                                        "addresses": [
                                            {
                                                "id": 11,
                                                "label": "Rumah"
                                            }
                                        ],
                                        "saldo": 500000,
                                        "midtrans": {
                                            "env": "sandbox",
                                            "client_key": "SB-Mid-client-xxx"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tidak terautentikasi."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/checkout/midtrans/token": {
            "post": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Buat Midtrans Snap token untuk checkout",
                "operationId": "1c8f2f1a702dafecc6b6a343b611c3c7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "address_mode",
                                    "order_type",
                                    "shipping_cost"
                                ],
                                "properties": {
                                    "address_mode": {
                                        "type": "string",
                                        "enum": [
                                            "saved",
                                            "manual"
                                        ],
                                        "example": "saved"
                                    },
                                    "order_type": {
                                        "type": "string",
                                        "enum": [
                                            "planA",
                                            "planB"
                                        ],
                                        "example": "planA"
                                    },
                                    "address_id": {
                                        "type": "integer",
                                        "example": 12,
                                        "nullable": true
                                    },
                                    "recipient_name": {
                                        "type": "string",
                                        "example": "Budi Santoso",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "08123456789",
                                        "nullable": true
                                    },
                                    "address_line": {
                                        "type": "string",
                                        "example": "Jl. Merdeka No. 1, RT 02/RW 01",
                                        "nullable": true
                                    },
                                    "province": {
                                        "type": "string",
                                        "example": "JAWA BARAT",
                                        "nullable": true
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "KOTA BANDUNG",
                                        "nullable": true
                                    },
                                    "province_id": {
                                        "type": "integer",
                                        "example": 9,
                                        "nullable": true
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "example": 501,
                                        "nullable": true
                                    },
                                    "district": {
                                        "type": "string",
                                        "example": "COBLONG",
                                        "nullable": true
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "40132",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Tolong kirim sore hari.",
                                        "nullable": true
                                    },
                                    "shipping_service_code": {
                                        "type": "string",
                                        "example": "REGPACK",
                                        "nullable": true
                                    },
                                    "shipping_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15000
                                    },
                                    "shipping_etd": {
                                        "type": "string",
                                        "example": "2-3 Hari",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Snap token berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "snapToken": "midtrans-snap-token",
                                    "orderId": 1201,
                                    "orderNo": "ORD-20260301-ABC123"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tidak terautentikasi."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Data tidak valid.",
                                    "errors": {
                                        "field": [
                                            "Field wajib diisi."
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Gagal membuat token",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Gagal membuat token pembayaran Midtrans."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dengan format: Bearer {token}",
                "bearerFormat": "Bearer Token",
                "scheme": "bearer"
            }
        }
    },
...
}